Remove explicit types where not needed

+++ src/no/runsafe/toybox/command/Nuke.java
@@ -25,12 +25,12 @@ public class Nuke extends PlayerCommand
 	@Override
 	public String OnExecute(IPlayer executor, IArgumentList parameters)
 	{
-		int radius = (Integer) parameters.getValue(RADIUS);
+		int radius = parameters.getValue(RADIUS);
 
 		if (radius > 15)
 			return "Max radius size: 15";
 
-		int power = (Integer) parameters.getValue(POWER);
+		int power = parameters.getValue(POWER);
 		ILocation location = executor.getLocation();
 
 		location.decrementX(radius);
